use an array for passing many options
In zsh a for loops iterates over an array. For bash you would use a
space-separated string without double quotes like

v="a b c"
for item in $v  # no double quotes!

In zsh this would iterate once with the item "a b c", like bash would do
with "$v". But there are arrays!