use real unicode in mysql
Annoyingly, by default MySQL uses what they call "utf8", which is
BMP-only, up-to-3-bytes utf8.

You have to manually choose "utf8mb4" which means up-to-4-bytes,
all-planes utf8.

Then you need to set it for the mysql client as well, because for some
reason it doesn't detect the server charset.