Use full vim on both macos and fedora
EDITOR is the default system editor used for example on git commit. We
want to use a nice, full featured vim, and not some older, limited
version that is also installed. I alias the good vim to "vi", but EDITOR
doesn't follow aliases.

On macos, we install a recent and complete vim with `brew install vim`.
Since I don't install it with the override-system-vi version, I need to
call it with "vim".

On Linux, things are confusing as usual. On both Fedora and Ubuntu the
good, complete vim that supports the system clipboard is the GUI
version, that can be used in a terminal. Hence "vimx".