Installing stuff doesn't belong in vimrc
A plugin manager is actually not necessary. Vim can now manage the
runtimepath with the packages feature, so they just need to be fetched
and they will work.

Before the packages feature existed in vim, pathogen was a good
replacement. Fetching the plugins can be done somewhere else, like with
ansible for example.